+++ firmware/hls/topotrigger/NN/parameters.h
@@ -1,0 +1,95 @@
+#ifndef PARAMETERS_H_
+#define PARAMETERS_H_
+
+#include "ap_int.h"
+#include "ap_fixed.h"
+
+#include "nnet_utils/nnet_helpers.h"
+//hls-fpga-machine-learning insert includes
+#include "nnet_utils/nnet_activation.h"
+#include "nnet_utils/nnet_activation_stream.h"
+#include "nnet_utils/nnet_dense.h"
+#include "nnet_utils/nnet_dense_compressed.h"
+#include "nnet_utils/nnet_dense_stream.h"
+ 
+//hls-fpga-machine-learning insert weights
+#include "weights/w2.h"
+#include "weights/b2.h"
+#include "weights/w5.h"
+#include "weights/b5.h"
+
+//hls-fpga-machine-learning insert layer-config
+// newfwkbothchannel_small_compressed0_fc1
+struct config2 : nnet::dense_config {
+    static const unsigned n_in = N_INPUT_1_1;
+    static const unsigned n_out = N_LAYER_2;
+    static const unsigned io_type = nnet::io_parallel;
+    static const unsigned strategy = nnet::latency;
+    static const unsigned reuse_factor = 1;
+    static const unsigned n_zeros = 1339;
+    static const unsigned n_nonzeros = 325;
+    static const bool store_weights_in_bram = false;
+    typedef ap_fixed<16,6> accum_t;
+    typedef bias2_t bias_t;
+    typedef weight2_t weight_t;
+    typedef ap_uint<1> index_t;
+    template<class x_T, class y_T, class res_T>
+    using product = nnet::product::mult<x_T, y_T, res_T>;
+};
+
+// newfwkbothchannel_small_compressed0_fc1_linear
+struct linear_config3 : nnet::activ_config {
+    static const unsigned n_in = N_LAYER_2;
+    static const unsigned table_size = 1024;
+    static const unsigned io_type = nnet::io_parallel;
+    static const unsigned reuse_factor = 1;
+    typedef ap_fixed<18,8> table_t;
+};
+
+// newfwkbothchannel_small_compressed0_activation1
+struct relu_config4 : nnet::activ_config {
+    static const unsigned n_in = N_LAYER_2;
+    static const unsigned table_size = 1024;
+    static const unsigned io_type = nnet::io_parallel;
+    static const unsigned reuse_factor = 1;
+    typedef ap_fixed<18,8> table_t;
+};
+
+// newfwkbothchannel_small_compressed0_output
+struct config5 : nnet::dense_config {
+    static const unsigned n_in = N_LAYER_2;
+    static const unsigned n_out = N_LAYER_5;
+    static const unsigned io_type = nnet::io_parallel;
+    static const unsigned strategy = nnet::latency;
+    static const unsigned reuse_factor = 1;
+    static const unsigned n_zeros = 51;
+    static const unsigned n_nonzeros = 13;
+    static const bool store_weights_in_bram = false;
+    typedef ap_fixed<16,6> accum_t;
+    typedef bias5_t bias_t;
+    typedef weight5_t weight_t;
+    typedef ap_uint<1> index_t;
+    template<class x_T, class y_T, class res_T>
+    using product = nnet::product::mult<x_T, y_T, res_T>;
+};
+
+// newfwkbothchannel_small_compressed0_output_linear
+struct linear_config6 : nnet::activ_config {
+    static const unsigned n_in = N_LAYER_5;
+    static const unsigned table_size = 1024;
+    static const unsigned io_type = nnet::io_parallel;
+    static const unsigned reuse_factor = 1;
+    typedef ap_fixed<18,8> table_t;
+};
+
+// newfwkbothchannel_small_compressed0_sigmoid
+struct sigmoid_config7 : nnet::activ_config {
+    static const unsigned n_in = N_LAYER_5;
+    static const unsigned table_size = 1024;
+    static const unsigned io_type = nnet::io_parallel;
+    static const unsigned reuse_factor = 1;
+    typedef ap_fixed<18,8> table_t;
+};
+
+
+#endif